Moved VaultTestKekManager to the AbstractVaultTestKmsFacade
